Level Visuals
Added RobotLab to the project, but without the settings. If import settings it screws up connection to vive.  Using empty lab prefab as the starting point of the level. Replaced the spawning points with teleportpad mesh, then changed the default particle system settings to get a different sort of look. Enemies now has a 1 second delay for spawning to give the teleporter animation a head start.